Applied a myriad of minor linter fixes. All tests and benches passing.

+++ deep_causality/benches/benchmarks/bench_linear_graph.rs
@@ -23,7 +23,7 @@ fn small_linear_graph_benchmark(criterion: &mut Criterion)
         )
     });
 
-    let x = (data.len() / 2) as usize;
+    let x = data.len() / 2;
     let index = NodeIndex::new(x);
 
     criterion.bench_function("small_linear_graph_reason_subgraph_from_cause", |bencher| {
@@ -32,7 +32,7 @@ fn small_linear_graph_benchmark(criterion: &mut Criterion)
         )
     });
 
-    let x = (data.len() / 2) as usize;
+    let x = data.len() / 2;
     let start_index = NodeIndex::new(x);
     let stop_index = NodeIndex::new(x + 25);
 
@@ -61,7 +61,7 @@ fn medium_linear_graph_benchmark(criterion: &mut Criterion)
         )
     });
 
-    let x = (data.len() / 2) as usize;
+    let x = data.len() / 2;
     let index = NodeIndex::new(x);
 
     criterion.bench_function("medium_linear_graph_reason_subgraph_from_cause", |bencher| {
@@ -70,7 +70,7 @@ fn medium_linear_graph_benchmark(criterion: &mut Criterion)
         )
     });
 
-    let x = (data.len() / 2) as usize;
+    let x = data.len() / 2;
     let start_index = NodeIndex::new(x);
     let stop_index = NodeIndex::new(x + 25);
 
@@ -99,7 +99,7 @@ fn large_linear_graph_benchmark(criterion: &mut Criterion)
         )
     });
 
-    let x = (data.len() / 2) as usize;
+    let x = data.len() / 2;
     let index = NodeIndex::new(x);
 
     criterion.bench_function("large_linear_graph_reason_subgraph_from_cause", |bencher| {
@@ -108,7 +108,7 @@ fn large_linear_graph_benchmark(criterion: &mut Criterion)
         )
     });
 
-    let x = (data.len() / 2) as usize;
+    let x = data.len() / 2;
     let start_index = NodeIndex::new(x);
     let stop_index = NodeIndex::new(x + 25);
